Integrate patch modules into core
