+++ nicegold_v5/AGENTS.md
@@ -322,3 +322,6 @@
 ### 2025-08-31
 - เพิ่มฟังก์ชัน safe_calculate_net_change สำหรับคำนวณกำไรสุทธิอย่างปลอดภัย (Patch I)
 
+### 2025-09-01
+- รวม patch_v11x และ patch_i_tp1tp2_fix เข้ากับโมดูลหลัก และลบไฟล์ patch
+
