--- patch_v11x.py
@@ -1,3 +1,0 @@
-from nicegold_v5.entry import simulate_trades_with_tp
-
-__all__ = ["simulate_trades_with_tp"]
